+++ r-cfdnalab/R/helpers.R
@@ -7,7 +7,7 @@
 CFDNALAB_SCHEMA_VERSION_RANGES <- list(
   midpoint_profiles = c(min = 1L, max = 1L),
   end_motif_counts = c(min = 1L, max = 2L),
-  ref_kmer_frequencies = c(min = 1L, max = 1L)
+  ref_kmer_frequencies = c(min = 2L, max = 2L)
 )
 
 #' Validate a cfDNAlab Zarr store path.